+++ apps/cli/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @pipelab/cli
 
+## 2.0.1-latest.38
+
+### Patch Changes
+
+- qsd
+- Updated dependencies
+  - @pipelab/constants@1.0.1-latest.35
+  - @pipelab/core-node@1.0.1-latest.39
+  - @pipelab/shared@2.0.1-latest.36
+
 ## 2.0.1-latest.37
 
 ### Patch Changes