+++ assets/asset-electron/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pipelab/asset-electron
 
+## 1.0.1-latest.38
+
+### Patch Changes
+
+- qsd
+
 ## 1.0.1-latest.37
 
 ### Patch Changes