+++ plugins/plugin-filesystem/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pipelab/plugin-filesystem
 
+## 1.0.1-latest.30
+
+### Patch Changes
+
+- qsd
+- Updated dependencies
+  - @pipelab/plugin-core@1.0.1-latest.35
+
 ## 1.0.1-latest.29
 
 ### Patch Changes